Move most H5P injection functions into new module
...instead of nesting them inside one function

- Flat is better than nested.
- Sparse is better than dense.
- Readability counts.
